Restructure files (#1560)
* Move handler tag generation and id to tag map to separate file

* Move allowed props to separate file

* Rename gestureHandlers to gestureHandlerCommon

* Split gestureHandlers.ts into separate files for every handler

* Update comment
