+++ src/handlers/handlersRegistry.ts
@@ -1,0 +1,7 @@
+export const handlerIDToTag: Record<string, number> = {};
+
+let handlerTag = 1;
+
+export function getNextHandlerTag(): number {
+  return handlerTag++;
+}
